Adding .obj to version control. Trying laptop
I don't get what is so hard about getting this to run on another machine. I'm an hour and 10 minutes into setting things up on the laptop. I think I need to actually save the vsproject and some of the settings to github so that it actually works out of the box.